Polynomials

Base.Math.exponent

Base.inv

MultivariateSeries.deg

MultivariateSeries.monoms

MultivariateSeries.sparse_pol

MultivariateSeries.@ring

MultivariateSeries.@ring — Macro

@ring args...

Defines the arguments as variables and output their array.

Example

X = @ring x1 x2

MultivariateSeries.monoms — Function

monoms(V, d::Int64) -> Vector{Monomial}
monoms(V, rg::UnitRangeInt64) -> Vector{Monomial}

List of all monomials in the variables V up to degree d of from degree d1 to d2, ordered by increasing degree.

monoms(V, d::Int64) -> Vector{Monomial}

List of all monomials in the variables V up to degree d of from degree d1 to d2, ordered by increasing degree.

Base.inv — Function

 inv(m :: Monomial{true})

return the inverse monomial with opposite exponents.

 inv(m :: Monomial)

return the inverse monomial with opposite exponents.

Base.Math.exponent — Function

exponent(m::Monomial) -> Array{Int64,1}

Get the exponent of a monomial as an array of Int64

MultivariateSeries.deg — Function

deg(p:Polynomial) -> Int64

Degree of a polynomial

MultivariateSeries.sparse_pol — Function

sparse_pol(w, E, X) -> Polynomial{true,C}

Compute the polynomial $∑ ωᵢ X^E[i,:]$ with coefficients $ωᵢ$ and monomial exponents $E[i,:]$.