Functions for series

MultivariateSeries.prodset — Function

Product-wise vector of X by Y ordered by row with no repetition: if xiyj appears as xi'yj' with i'<i or i'=i and j'<j it is not inserted.

MultivariateSeries.prodvec — Function

Product-wise vector of X by Y ordered by row: [x1y1, x1y2, ..., x2*y1, ....]

MultivariateSeries.integrate — Function

Compute the primitive $\int_v s = v^{-1}*s$

Compute the truncated primitive ``\int{x{i}} s{| x{i+1}=0, ...x_n=0}''

MultivariateSeries.matrixof — Function

Coefficient matrix of the polynomials in P with respect to the monomial vector L